PP-1447 Fix for pull requests REBASED into master instead of MERGED
Github's new feature allowing to REBASE pull requests into master instead of MERGE
has the side effect to not create a merge commit which pay-admin-dashboad
relies upon. When no merge commit is found, simply ignore and show N/A
for all the information that cannot be pulled as a result of that.